chore: fix HMR issue
