+++ shared/src/containers/Feed/components/CommentInput/index.ts
@@ -1,0 +1,2 @@
+export * from './types'
+export { default } from './CommentInput'
