+++ shared/src/containers/Feed/components/CommentInput/types.ts
@@ -1,0 +1,17 @@
+export enum VersionReviewFeedback {
+  APPROVE = 'approve',
+  REQUEST_CHANGES = 'request_changes',
+}
+
+export const mentionTypeOptions = {
+  '@@@': {
+    id: 'task',
+  },
+  '@@': {
+    id: 'version',
+  },
+  '@': {
+    id: 'user',
+    isCircle: true,
+  },
+}
